+++ engine/flink/management/periodic/src/main/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicDeploymentManagerProvider.scala
@@ -6,13 +6,13 @@ import pl.touk.nussknacker.engine.api.TypeSpecificData
 import pl.touk.nussknacker.engine.api.deployment.DeploymentManager
 import pl.touk.nussknacker.engine.api.queryablestate.QueryableClient
 import pl.touk.nussknacker.engine.management.FlinkConfig
-import pl.touk.nussknacker.engine.management.periodic.service.{AdditionalDeploymentDataProvider, DefaultAdditionalDeploymentDataProvider, EmptyPeriodicProcessListenerFactory, PeriodicProcessListenerFactory}
+import pl.touk.nussknacker.engine.management.periodic.service.{AdditionalDeploymentDataProvider, DefaultAdditionalDeploymentDataProvider, ProcessConfigEnricherFactory, EmptyPeriodicProcessListenerFactory, PeriodicProcessListenerFactory}
 import pl.touk.nussknacker.engine.util.config.ConfigEnrichments.RichConfig
 import pl.touk.nussknacker.engine.{DeploymentManagerProvider, ModelData}
 
 class PeriodicDeploymentManagerProvider(delegate: DeploymentManagerProvider,
                                         schedulePropertyExtractorFactory: SchedulePropertyExtractorFactory = _ => CronSchedulePropertyExtractor(),
-                                        enrichDeploymentWithJarDataFactory: EnrichDeploymentWithJarDataFactory = EnrichDeploymentWithJarDataFactory.noOp,
+                                        processConfigEnricherFactory: ProcessConfigEnricherFactory = ProcessConfigEnricherFactory.noOp,
                                         listenerFactory: PeriodicProcessListenerFactory = EmptyPeriodicProcessListenerFactory,
                                         additionalDeploymentDataProvider: AdditionalDeploymentDataProvider = DefaultAdditionalDeploymentDataProvider
                                        ) extends DeploymentManagerProvider with LazyLogging {
@@ -30,7 +30,7 @@ class PeriodicDeploymentManagerProvider(delegate: DeploymentManagerProvider,
     PeriodicDeploymentManager(
       delegate = delegateDeploymentManager,
       schedulePropertyExtractorFactory = schedulePropertyExtractorFactory,
-      enrichDeploymentWithJarDataFactory = enrichDeploymentWithJarDataFactory,
+      processConfigEnricherFactory = processConfigEnricherFactory,
       periodicBatchConfig = periodicBatchConfig,
       flinkConfig = flinkConfig,
       originalConfig = config,