+++ engine/flink/management/periodic/src/test/scala/pl/touk/nussknacker/engine/management/periodic/JarManagerTest.scala
@@ -8,6 +8,7 @@ import org.scalatest.{FunSuite, Matchers}
 import pl.touk.nussknacker.engine.api.ProcessVersion
 import pl.touk.nussknacker.engine.api.process.ProcessName
 import pl.touk.nussknacker.engine.management.periodic.flink.FlinkJarManager
+import pl.touk.nussknacker.engine.management.periodic.service.ProcessConfigEnricher
 import pl.touk.nussknacker.engine.modelconfig.InputConfigDuringExecution
 import pl.touk.nussknacker.test.PatientScalaFutures
 
@@ -35,8 +36,7 @@ class JarManagerTest extends FunSuite
       flinkClient = new FlinkClientStub,
       jarsDir = jarsDir,
       modelConfig = InputConfigDuringExecution(ConfigFactory.empty()),
-      createCurrentModelJarFile = currentModelJarFile,
-      enrichDeploymentWithJarData = EnrichDeploymentWithJarData.noOp
+      createCurrentModelJarFile = currentModelJarFile
     )
   }